+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Release TBD
   * Separate ProductImage class into Images\Product.
   * Pass query string to product detail view for highlighting.
   * Add `View Cart` button to shop header when cart has items.
+  * Add shipping and tracking for orders.
 
 ## v0.7.1
 Release 2019-08-02